+++ x-pack/plugin/esql/compute/src/main/java/module-info.java
@@ -36,4 +36,5 @@ module org.elasticsearch.compute {
     exports org.elasticsearch.compute.aggregation.table;
     exports org.elasticsearch.compute.data.sort;
     exports org.elasticsearch.compute.querydsl.query;
+    exports org.elasticsearch.compute.lucene.read;
 }
